Friendbot, comments: Move services/horizon/internal/render/hal/main.go to support/render/hal/main.go
