+++ services/friendbot/internal/actions_friendbot.go
@@ -5,7 +5,7 @@ import (
 
 	client "github.com/stellar/go/clients/horizon"
 	"github.com/stellar/go/services/horizon/internal/actions"
-	"github.com/stellar/go/services/horizon/internal/render/hal"
+	"github.com/stellar/go/support/render/hal"
 	"github.com/stellar/go/services/horizon/internal/render/problem"
 	"github.com/zenazn/goji/web"
 )
