+++ services/horizon/internal/actions_data.go
@@ -2,7 +2,7 @@ package horizon
 
 import (
 	"github.com/stellar/go/services/horizon/internal/db2/core"
-	"github.com/stellar/go/services/horizon/internal/render/hal"
+	"github.com/stellar/go/support/render/hal"
 	"github.com/stellar/go/services/horizon/internal/render/sse"
 )
