+++ services/horizon/internal/actions_effects.go
@@ -6,9 +6,10 @@ import (
 
 	"github.com/stellar/go/services/horizon/internal/db2"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
-	"github.com/stellar/go/services/horizon/internal/resource"
 	"github.com/stellar/go/services/horizon/internal/render/hal"
 	"github.com/stellar/go/services/horizon/internal/render/sse"
+	"github.com/stellar/go/services/horizon/internal/resource"
+	halRender "github.com/stellar/go/support/render/hal"
 )
 
 // This file contains the actions:
@@ -41,7 +42,7 @@ func (action *EffectIndexAction) JSON() {
 	)
 
 	action.Do(func() {
-		hal.Render(action.W, action.Page)
+		halRender.Render(action.W, action.Page)
 	})
 }
