+++ services/horizon/internal/actions_ledger.go
@@ -4,10 +4,11 @@ import (
 	"github.com/stellar/go/services/horizon/internal/db2"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"github.com/stellar/go/services/horizon/internal/ledger"
-	"github.com/stellar/go/services/horizon/internal/resource"
 	"github.com/stellar/go/services/horizon/internal/render/hal"
 	"github.com/stellar/go/services/horizon/internal/render/problem"
 	"github.com/stellar/go/services/horizon/internal/render/sse"
+	"github.com/stellar/go/services/horizon/internal/resource"
+	halRender "github.com/stellar/go/support/render/hal"
 )
 
 // This file contains the actions:
@@ -32,7 +33,7 @@ func (action *LedgerIndexAction) JSON() {
 		action.ValidateCursorWithinHistory,
 		action.loadRecords,
 		action.loadPage,
-		func() { hal.Render(action.W, action.Page) },
+		func() { halRender.Render(action.W, action.Page) },
 	)
 }
 
@@ -100,7 +101,7 @@ func (action *LedgerShowAction) JSON() {
 		func() {
 			var res resource.Ledger
 			res.Populate(action.Ctx, action.Record)
-			hal.Render(action.W, res)
+			halRender.Render(action.W, res)
 		},
 	)
 }
