+++ services/horizon/internal/actions_root.go
@@ -3,7 +3,7 @@ package horizon
 import (
 	"github.com/stellar/go/services/horizon/internal/ledger"
 	"github.com/stellar/go/services/horizon/internal/resource"
-	"github.com/stellar/go/services/horizon/internal/render/hal"
+	"github.com/stellar/go/support/render/hal"
 )
 
 // RootAction provides a summary of the horizon instance and links to various
