+++ services/horizon/internal/actions_transaction.go
@@ -5,11 +5,12 @@ import (
 
 	"github.com/stellar/go/services/horizon/internal/db2"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
-	"github.com/stellar/go/services/horizon/internal/resource"
-	"github.com/stellar/go/services/horizon/internal/txsub"
 	"github.com/stellar/go/services/horizon/internal/render/hal"
 	"github.com/stellar/go/services/horizon/internal/render/problem"
 	"github.com/stellar/go/services/horizon/internal/render/sse"
+	"github.com/stellar/go/services/horizon/internal/resource"
+	"github.com/stellar/go/services/horizon/internal/txsub"
+	halRender "github.com/stellar/go/support/render/hal"
 )
 
 // This file contains the actions:
@@ -37,7 +38,7 @@ func (action *TransactionIndexAction) JSON() {
 		action.loadRecords,
 		action.loadPage,
 		func() {
-			hal.Render(action.W, action.Page)
+			halRender.Render(action.W, action.Page)
 		},
 	)
 }
@@ -126,7 +127,7 @@ func (action *TransactionShowAction) JSON() {
 		action.loadParams,
 		action.loadRecord,
 		action.loadResource,
-		func() { hal.Render(action.W, action.Resource) },
+		func() { halRender.Render(action.W, action.Resource) },
 	)
 }
 
@@ -147,7 +148,7 @@ func (action *TransactionCreateAction) JSON() {
 		action.loadResource,
 
 		func() {
-			hal.Render(action.W, action.Resource)
+			halRender.Render(action.W, action.Resource)
 		})
 }
